% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForExpressionSet.R
\name{checkForExpressionSet}
\alias{checkForExpressionSet}
\title{Check if object exists on the study server and if it's an ExpressionSet}
\usage{
checkForExpressionSet(set, datasources = NULL)
}
\arguments{
\item{set}{\code{character} Name of the object to look for on the study server}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
No return object, only stops the execution if the object is not on the study server or if it's not an
ExposomeSet
}
\description{
Internal function
}
