% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eafPlot.R
\name{.eafPlot}
\alias{.eafPlot}
\title{Auxiliary function for \code{eafPlot}}
\usage{
.eafPlot(
  x,
  freq_x,
  freq_reference,
  inverse_freq_x,
  inverse_freq_reference,
  title = NULL
)
}
\description{
Auxiliary function for \code{eafPlot}
}
