% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzPlot.R
\name{.pzPlotgg}
\alias{.pzPlotgg}
\title{Auxiliary function for \code{pzPlot}}
\usage{
.pzPlotgg(x_single, beta_column, se_column, pval_column, title = NULL)
}
\description{
Auxiliary function for \code{pzPlot}
}
