% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.addPhenoData.R
\name{ds.addPhenoData}
\alias{ds.addPhenoData}
\title{Title}
\usage{
ds.addPhenoData(x, pheno, ID = "ID", name = NULL, datasources = NULL)
}
\arguments{
\item{x}{}

\item{pheno}{}

\item{ID}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
