% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.cbind_unsafe.R
\name{ds.cbind_unsafe}
\alias{ds.cbind_unsafe}
\title{Title}
\usage{
ds.cbind_unsafe(obj1, obj2, name, datasources = NULL)
}
\arguments{
\item{obj1}{}

\item{obj2}{}

\item{name}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
