% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.createRSE.R
\name{ds.createRSE}
\alias{ds.createRSE}
\title{Create a RangedSummarizedExperiment from a RNAseq count table and 
a phenotypes table}
\usage{
ds.createRSE(rnaseq, phe, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{rnaseq}{\code{character} Name of the RNAseq table on the server}

\item{phe}{\code{character} Name of the phenotypes table on the server}

\item{newobj.name}{\code{character} (default \code{NULL}) Name to be assigned 
on the server to the RangedSummarizedExperiment. If \code{NULL} the created RSE will 
be assigned to a variable named \code{'RSE'}}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Create a RangedSummarizedExperiment from a RNAseq count table and 
a phenotypes table
}
\details{
The RNAseq count table has to have the following structure: \cr
- First column: Entrez identificator \cr
- Second column: Annotated symbol \cr
- All the remaining columns: One col. per individual \cr
The phenotypes table has to have the following structure: \cr
- First column: ID of the individuals. To match with the column names 
of the RNAseq table \cr
- All the remaining columns: One col. per phenotype
}
