% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DESeq2.R
\name{ds.DESeq2}
\alias{ds.DESeq2}
\title{Server-side Differential Gene Expression analysis using DESeq2}
\usage{
ds.DESeq2(
  model,
  set,
  test = "Wald",
  fitType = "parametric",
  sfType = "ratio",
  reduced = NULL,
  contrast = NULL,
  datasources = NULL
)
}
\arguments{
\item{model}{formula indicating the condition and other covariates to be ajusted}

\item{set}{a \code{RangedSummarizedExperiment} object.}

\item{test}{\code{"Wald"} or \code{"LRT"}.}

\item{fitType}{\code{"parametric"}, \code{"local"}, or \code{"mean"}.}

\item{sfType}{\code{"ratio"}, \code{"poscounts"}, or \code{"iterate"}.}

\item{reduced}{a reduced formula for \code{test="LRT"}}

\item{contrast}{the comparison to extract from the object to build the results table}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. 
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
Data frame containing the results of the Differential Gene Expression analysis
}
\description{
This function performs a non-disclosive
Differential Gene Expression Analysis using \code{DESeq2} package
from Bioconductor.
}
\details{
Implementation of Bioconductor's \code{DESeq2} in DataSHIELD 
infraestructure. 

This function is similar to Bioconductor's 
\code{DESeq2} package \code{DESeq} function.
}
\author{
L. Abarrategui, for DataSHIELD development team
}
