% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.edgeR.R
\name{ds.edgeR}
\alias{ds.edgeR}
\title{Server-side Differential Gene Expression analysis using DESeq2}
\usage{
ds.edgeR(
  model,
  set,
  test = "QLF",
  dispersion = "both",
  levels = "design",
  coef = 2,
  contrast = NULL,
  datasources = NULL
)
}
\arguments{
\item{model}{formula indicating the condition and other covariates to be ajusted}

\item{set}{a}

\item{test}{}

\item{contrast}{the comparison to extract from the object to build the results table}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. 
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
Data frame containing the results of the Differential Gene Expression analysis
}
\description{
This function performs a non-disclosive
Differential Gene Expression Analysis using \code{DESeq2} package
from Bioconductor.
}
\details{
Implementation of Bioconductor's \code{DESeq2} in DataSHIELD 
infraestructure. 

This function is similar to Bioconductor's 
\code{DESeq2} package \code{DESeq} function.
}
\author{
L. Abarrategui, for DataSHIELD development team
}
