% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.featureData.R
\name{ds.featureData}
\alias{ds.featureData}
\title{Retrieve information on features from eSets and RSEs.}
\usage{
ds.featureData(object, datasources = NULL)
}
\arguments{
\item{object}{name of the DataSHIELD object to which the ExpresionSet has been assigned}

\item{datasources}{....}
}
\value{
\code{featureNames} returns a (usually long!) character vector uniquely identifying each feature.
}
\description{
This function is similar to Bioconductor function \code{featureData}
}
\examples{

}
