% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.filterByExpr.R
\name{ds.filterByExpr}
\alias{ds.filterByExpr}
\title{Filter Genes By Expression Level}
\usage{
ds.filterByExpr(object, datasources = NULL)
}
\arguments{
\item{object}{name of the DataSHIELD object to which the ExpresionSet, SummarizedExperiment or RangedSummarizedExperiment has been assigned}

\item{datasources}{....}
}
\value{
initial object having only genes with sufficiently large counts to be retained in a statistical analysis.
}
\description{
This function is similar to Bioconductor function \code{edgeR::filterByExpr}
}
\examples{

}
