% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.fvarLabels.R
\name{ds.fvarLabels}
\alias{ds.fvarLabels}
\title{Retrieve phenotypic data (e.g. covariates) from eSets and RSEs.}
\usage{
ds.fvarLabels(object, datasources = NULL)
}
\arguments{
\item{object}{name of the DataSHIELD object to which the ExpresionSet has been assigned}

\item{datasources}{....}
}
\value{
a character vector of measured variables.
}
\description{
This function is similar to Bioconductor function \code{varLabels}
}
\examples{

}
