% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GenotypeData.R
\name{ds.GenotypeData}
\alias{ds.GenotypeData}
\title{Creates a GenotypeData object at each server}
\usage{
ds.GenotypeData(x, covars, columnId, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{x}{a \code{GdsGenotypeReader} object (see GWASTools). It is the object that is obtained in the opal servers
from a resource of type VCF2GDS}

\item{covars}{a data.frame or a tibble having the metadata of samples (i.e. phenotypes and/or covariates)}

\item{datasources}{a list of \code{DSConnection-class} objects obtained after login. If the <datasources> the default set of connections will be used: see \code{datashield.connections_default}.}

\item{data}{name of the DataSHIELD object to which the genotype (snpMatrix) and phenotypic data (data.frame) has been assigned}
}
\description{
The GenotypeData class is a container for storing genotype data from a GWAS 
together with the metadata associated with the subjects (i.e. phenotypes and covariates) and SNPs involved in the study.
}
\examples{

}
