% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.getChromosomeNames.R
\name{ds.getChromosomeNames}
\alias{ds.getChromosomeNames}
\title{Get names of chromosomes}
\usage{
ds.getChromosomeNames(genoData, datasources = NULL)
}
\arguments{
\item{genoData}{\code{character} Name of the \code{\link{GenotypeData}} object on the server}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
A \code{list} with: \cr
-autosomes: integer codes for the autosomes \cr
-Xchr: integer code for the X chromosome \cr
-pseudoautosomalRegionXY: integer code for the pseudoautosomal region \cr
-Ychr: integer code for the Y chromosome \cr
-mitochondrial: integer code for mitochondrial SNPs \cr
}
\description{
Get the names of the chromosomes of a GenotypeData object on the server side
}
