% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.getSNPs.R
\name{ds.getSNPs}
\alias{ds.getSNPs}
\title{Retrieve SNPs from}
\usage{
ds.getSNPs(genoData, datasources = NULL)
}
\arguments{
\item{genoData}{\code{character} Object of class \code{GenotypeData} or 
\code{GdsGenotypeReader} on the server}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Character vector with the SNP names
}
\description{
Retrieve SNPs from
}
