% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.getSNPSbyGen.R
\name{ds.getSNPSbyGen}
\alias{ds.getSNPSbyGen}
\title{Title}
\usage{
ds.getSNPSbyGen(gdsFile, genes, name = NULL, datasources = NULL)
}
\arguments{
\item{gdsFile}{}

\item{genes}{}

\item{name}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
