% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.glmSNP.R
\name{ds.glmSNP}
\alias{ds.glmSNP}
\title{Logistic regression analysis of pooled data for each SNP site in study}
\usage{
ds.glmSNP(
  snps.fit = NULL,
  model,
  genoData,
  datasources = NULL,
  type.p.adj = "fdr",
  mc.cores = 1,
  family = "binomial",
  strata = NULL
)
}
\arguments{
\item{snps.fit}{an optional parameter input as a character vector of
SNPs (rs numbers) that should be analysed. If missing all SNPs are analysed}

\item{model}{list of phenotypic variables to use as covariates in the regression analysis in the form:
"outcome ~ covar1 + covar2 +  ... + covarN"}

\item{genoData}{name of the DataSHIELD object to which the genotype (snpMatrix) and phenotypic data (data.frame) has been assigned}

\item{datasources}{Opal object or list of opal objects denoting the opal server(s) information}

\item{mc.cores}{optional parameter that allows the user to specify the number of CPU cores to use}

\item{strata}{\code{character} Categorical variable to perform a stratified glm}
}
\description{
Fits a generalized linear model to gentic data for each SNP
in the data sets considered, using user specified outcome and phenotypic variabes as covariates
Outputs a matrix containing a beta value, standard error and p-value for each SNP
}
\examples{

}
