% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.index.gdsn.R
\name{ds.index.gdsn}
\alias{ds.index.gdsn}
\title{Retrieve phenotypic data (e.g. covariates) from eSets.}
\usage{
ds.index.gdsn(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{an object of class}

\item{datasources}{....}
}
\value{
a character vector of measured variables.
}
\description{
This function is similar to function \code{index.gdsn}  in the 'gdsfmt' package
}
