% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lmFeature.R
\name{ds.lmFeature}
\alias{ds.lmFeature}
\title{Linear regression analysis of pooled data for each CpG site in study}
\usage{
ds.lmFeature(features = NULL, model, eSet, connections = NULL,
  type.p.adj = "fdr", cellCountsAdjust = FALSE, mc.cores = 1)
}
\arguments{
\item{features}{an optional parameter input as a vector of integer values which indicates the indices of specific 
features (e.g. genes, CpGs, ...) that should be analysed. If missing all features are analysed}

\item{model}{formula indicating the condition (left side) and other covariates to be adjusted for 
(i.e. condition ~ covar1 + ... + covar2). The fitted model is: feature ~ condition + covar1 + ... + covarN}

\item{eSet}{name of the DataSHIELD object to which the ExpresionSet has been assigned}

\item{connections}{....}

\item{type.p.adj}{multiple comparison correction method. Default 'fdr'}

\item{cellCountsAdjust}{logical value which indicates whether models should be
adjusted for cell counts that are estimated using 'meffil.estimate.cell.counts.from.betas'
function from \code{meffil} package.}

\item{mc.cores}{optional parameter that allows the user to specify the number of CPU cores to use during 
parallel processing. Argument can only be > 1 when the function is run on a linux machine
models should be adjusted for the estimated cell counts by including the variables in the models.
NOTE: This assumes that the Opal pheno tables for every study include the necessary estimated cell count data 
originally computed when running the createOpalFiles function}
}
\description{
Performing a linear regression analysis on pooled data from multiple studies for every feature
}
\details{
The function fits a generalized linear model of a ExpressionSet for each feature (gene, CpG site, ...) 
in the data sets considered, using user specified condition and covariates
Outputs a matrix containing a beta value, standard error and p-value for each feature
}
\examples{

}
