% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.make_valid_names.R
\name{ds.make_valid_names}
\alias{ds.make_valid_names}
\title{Rename column names and character variables}
\usage{
ds.make_valid_names(x, newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the table on the server to be modified}

\item{newobj}{\code{character} (default \code{NULL}) Name of the new object on the server. If \code{NULL} 
the input object \code{x} will be overwritten}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. 
If the \code{datasources} argument is not specified}
}
\description{
Passes the column names and the character variables through \code{make.names}, this creates 
names without any special characters so there are no problems for accessing any columns or any category 
when using the DataSHIELD parser.
}
