% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.metaGWAS.R
\name{ds.metaGWAS}
\alias{ds.metaGWAS}
\title{Genome-wide association analysis (GWAS)}
\usage{
ds.metaGWAS(
  genoData,
  model,
  family = "binomial",
  snpBlock = 10000,
  datasources = NULL,
  ...
)
}
\arguments{
\item{genoData}{\code{character vector} of objects on the server side object which is a container for storing genotype data
from a GWAS toghether with the metadata associated with the subjects (i.e. phenotypes and/or covariates)}

\item{model}{formula indicating the condition (left side) and other covariates to be adjusted for 
(i.e. condition ~ covar1 + ... + covar2). The fitted model is: snp ~ condition + covar1 + ... + covarN}

\item{family}{A description of the generalized linear model used. The defatul is "binomial" that is defined
for case/control studies. Quantitative traits can be analyzed by using "gaussian". Other values are accepted.}

\item{snpBlock}{an integer specifying the number of SNPs in an iteration block. See \code{GenotypeIterator} 
function in GWASTools package.}

\item{...}{other arguments of \code{fitNullModel} function in GENESIS package}
}
\value{
a matrix with SNPs ordered by p-values
}
\description{
Performs GWAS using GENESIS
}
\author{
Gonzalez, JR.
}
