% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.methylation_array_convert.R
\name{ds.methylation_array_convert}
\alias{ds.methylation_array_convert}
\title{Convert a Methylation ExpressionSet from EPIC to 450k Array (DataSHIELD Client Function)}
\usage{
ds.methylation_array_convert(
  eSet_symbol,
  objective_array,
  new.obj = NULL,
  datasources = NULL
)
}
\arguments{
\item{eSet_symbol}{A character string representing the server-side symbol of the ExpressionSet object containing either EPIC or 450k array data.}

\item{objective_array}{A character string specifying the target array type. Can be either "450k" or "epic".}

\item{new.obj}{A character string specifying the name of the new server-side object created. If NULL (default), the same name as `eSet_symbol` is used.}

\item{datasources}{A DataSHIELD connections object containing the information for one or multiple data sources.}
}
\value{
A server-side symbol representing the ExpressionSet object containing the subset of probes based on the target array type.
}
\description{
This function subsets a methylation ExpressionSet from an EPIC array (850k probes) to a 450k array or vice versa in a DataSHIELD environment.
}
\examples{
# Assuming you have a DataSHIELD login and 'eSet' is available on the server
# Convert the eSet from EPIC array to 450k array
eSet_450k_symbol <- ds.methylation_array_convert(eSet_symbol = "eSet", objective_array = "450k")
# Convert the eSet from 450k array to EPIC array
eSet_epic_symbol <- ds.methylation_array_convert(eSet_symbol = "eSet", objective_array = "epic")
}
