% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.nSamples.R
\name{ds.nSamples}
\alias{ds.nSamples}
\title{Retrieve the number of samples for eSets and RSEs.}
\usage{
ds.nSamples(object, datasources = NULL)
}
\arguments{
\item{object}{name of the DataSHIELD object to which the ExpresionSet or RangedSummarizedExperiment has been assigned}

\item{datasources}{....}
}
\value{
a numeric value
}
\description{
This function is similar to Bioconductor function \code{dim}
}
\examples{

}
