% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.PCASNPS.R
\name{ds.PCASNPS}
\alias{ds.PCASNPS}
\title{Title}
\usage{
ds.PCASNPS(gds, prune = TRUE, ld.threshold = 0.2, datasources = NULL)
}
\arguments{
\item{gds}{}

\item{ld.threshold}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
