% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.pData.R
\name{ds.pData}
\alias{ds.pData}
\title{Retrieve information on experimental phenotypes recorded in eSet and ExpressionSet-derived classes}
\usage{
ds.pData(x, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of an object (on the study server), 
possibly derived from \link{eSet-class} or \link{AnnotatedDataFrame}}

\item{newobj.name}{\code{character} (default \code{NULL}) If \code{NULL}, the created object will take the 
name \code{"new_pData"}}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data.frame on the study server.
}
\description{
These generic functions access the phenotypic data (e.g., covariates) 
and meta-data (e.g., descriptions of covariates) associated with an experiment.
}
