% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.PLINK.R
\name{ds.PLINK}
\alias{ds.PLINK}
\title{Genome-wide association analysis (GWAS) using PLINK}
\usage{
ds.PLINK(client, plink.arguments, datasources = NULL)
}
\arguments{
\item{client}{...}

\item{plink.command}{...}

\item{datasourcers}{...}
}
\value{
...
}
\description{
Performs GWAS using PLINK using shell command lines
}
\author{
Gonzalez, JR.
}
