% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.removeOutliers.R
\name{ds.removeOutliers}
\alias{ds.removeOutliers}
\title{Filter potential CpG outliers}
\usage{
ds.removeOutliers(object, pct = 0.125, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{object}{\code{character} Name of an \code{ExpressionSet} or \code{SummarizedExperiment / RangedSummarizedExperiment}
on the server side to which perform the filtering}

\item{pct}{\code{numeric} (default \code{0.125}) Tail and head quantiles that will be considered as 
outliers; for example: a \code{pct=0.125} will use the function 
\code{matrixStats::rowQuantiles(probs = c(0.125, 1-0.125))} to detect outliers.}

\item{newobj.name}{\code{character} (default \code{NULL}) If provided, the filtered object will be assigned
to a variable on the server with this name. If \code{NULL}, the input variable will be overwritten.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
This function does not have an output. It creates (or overwrites) an \code{ExpressionSet} on the study server.
}
\description{
Trimming methylation beta values to remove potential outliers.
Trimming scheme is as follows: trim values beyond the lower and upper outer fences. These are defined by 
the `pct` argument.
}
