% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.snptest.R
\name{ds.snptest}
\alias{ds.snptest}
\title{Interface to run SNPtest commands on a ssh connection (with SNPtest 2.5.2 installed)}
\usage{
ds.snptest(client, snptest.arguments, datasources = NULL)
}
\arguments{
\item{client}{\code{character} Name of the ssh resource on the server}

\item{snptest.arguments}{\code{character} SNPtest arguments, the leading "snptest" and trailing 
"-o output.out" do not need to be provided, they are handled on the server side.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
List containing: \cr
- Results: Table of results (typical SNPtest output file) \cr
- Output: Console output of the ssh query
}
\description{
Interface to run SNPtest commands on a ssh connection (with SNPtest 2.5.2 installed)
}
