% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.subsetExpressionSet.R
\name{ds.subsetExpressionSet}
\alias{ds.subsetExpressionSet}
\title{Subset ExpressionSet}
\usage{
ds.subsetExpressionSet(
  eSet,
  objective_variable = NULL,
  objective_value = NULL,
  complete_cases = FALSE,
  newobj.name = NULL,
  datasources = NULL
)
}
\arguments{
\item{eSet}{\code{character} ExpressionSet to subset}

\item{objective_variable}{\code{character} (default \code{NULL}) Name of the covariate on the ExpressionSet to use as filter}

\item{objective_value}{\code{character} (default \code{NULL}) Name of the value from the \code{objective_variable} to filter. 
The resulting subset will be the individuals that match this value. 
To put in in code, it can be represented as: \code{subset <- expressionSet[expressionSet$objective_variable == objective_value,]}}

\item{complete_cases}{\code{bool} (default \code{FALSE}) If \code{TRUE} only the complete cases will be included on the subset. 
This option can be used with \code{objective_variable} and \code{objective_value} or without them, if those arguments are not 
present, the subset will be the complete cases of the whole ExpressionSet.}

\item{newobj.name}{\code{character} (default \code{NULL}) Name of the subseted ExpressionSet. If \code{NULL}, the variable 
\code{"subsetted_ExpressionSet"} will be used.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Subset ExpressionSet using a categorical variable of the covariates as filter. Can also subset by 
complete cases
}
