% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tablePheno.R
\name{ds.tablePheno}
\alias{ds.tablePheno}
\title{Retrieve the number of counts at each factor levels for metadata covariates in eSets and RSEs.}
\usage{
ds.tablePheno(object, datasources = NULL)
}
\arguments{
\item{object}{name of the DataSHIELD object to which the ExpresionSet or RangedSummarizedExperiment has been assigned}

\item{datasources}{....}
}
\value{
a numeric value
}
\description{
This function is similar to base function \code{table}
}
\examples{

}
