% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eafPlot.R
\name{eafPlot}
\alias{eafPlot}
\title{EAF plot}
\usage{
eafPlot(
  x,
  reference,
  rs_x = "rs",
  rs_reference,
  freq_x = "freq",
  freq_reference,
  inverse_freq_x = T,
  inverse_freq_reference = F
)
}
\arguments{
\item{x}{\code{data.frame / list of data.frames (output of ds.GWAS)} Single table of results or list of tables 
(default output of \code{ds.GWAS}).}

\item{reference}{\code{data.frame} Reference table that at least contains: SNP rs IDs and minor (or major) 
allele frequency of reference.}

\item{rs_x}{\code{character} (default \code{"rs"}) Name of the column that contains the rs IDs on \code{x}}

\item{rs_reference}{\code{character} Name of the column that contains the rs IDs on \code{reference}}

\item{freq_x}{\code{character} (default \code{"freq"}) Name of the column that contains the minor (or major) 
allele frequencies on \code{x}}

\item{freq_reference}{\code{character} Name of the column that contains the minor (or major) 
allele frequencies on \code{reference}}

\item{inverse_freq_x}{\code{bool} (default \code{TRUE}) Use \code{TRUE} if \code{x} contains major allele 
frequencies, use \code{FALSE} if \code{x} contains minor allele frequencies. The default is set to 
\code{TRUE} since the default output of \code{ds.GWAS} frequencies is major allele frequencies.}

\item{inverse_freq_reference}{\code{bool} (default \code{FALSE}) Use \code{TRUE} if \code{x} contains major allele 
frequencies, use \code{FALSE} if \code{x} contains minor allele frequencies. The default is set to 
\code{TRUE} since usually reference panels report minor allele frequencies.}
}
\value{
A ggplot object
}
\description{
EAF plot
}
\details{
Plotting the reported EAFs against a reference set, such as from the HapMap or 
1000 Genomes projects, or from one specific study, can help visualize patterns that 
pinpoint strand issues, allele miscoding or the inclusion of individuals whose self-reported 
ancestry did not match their genetic ancestry.
}
