% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EGA_url.R
\name{get_EGA_url}
\alias{get_EGA_url}
\title{Get EGA URL}
\usage{
get_EGA_url(id, chr, start, end)
}
\arguments{
\item{id}{\code{character} ID of the file}

\item{chr}{\code{character} Chromosome to filter (encoding is server dependant)}

\item{start}{\code{numeric} Start position to filter}

\item{end}{\code{numeric} End position to filter}
}
\value{
URL of ga4gh VCF/BAM file
}
\description{
Get EGA URL
}
