% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ga4gh_url.R
\name{get_ga4gh_url}
\alias{get_ga4gh_url}
\title{Get GA4GH URL}
\usage{
get_ga4gh_url(url, id, chr, start, end, type)
}
\arguments{
\item{url}{\code{character} URL of the server}

\item{id}{\code{character} ID of the file}

\item{chr}{\code{character} Chromosome to filter (encoding is server dependant)}

\item{start}{\code{numeric} Start position to filter}

\item{end}{\code{numeric} End position to filter}

\item{type}{\code{character} Type of file (VCF/BAM)}
}
\value{
URL of ga4gh VCF/BAM file
}
\description{
Get GA4GH URL
}
