% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSNP.R
\name{glmSNP}
\alias{glmSNP}
\title{Fits a GLM to assess association between the outcome (binary variable) and a given SNP}
\usage{
glmSNP(snp, snps, genoData, vars, datasources, family = "binomial")
}
\value{
a vector with effect estimates, standard error and associated p-value
}
\description{
To be supplied
}
\author{
Gonzalez, JR.
}
