% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaNPlot.R
\name{lambdaNPlot}
\alias{lambdaNPlot}
\title{Lambda-N plot}
\usage{
lambdaNPlot(x, pval_column = "p.value", n_column = "n.obs", threshold = 1.2)
}
\arguments{
\item{x}{\code{list of data.frames (output of ds.GWAS)} List of tables (default output of \code{ds.GWAS}).}

\item{pval_column}{\code{character} (default \code{"p.value"}) Name of the column containing the p-values.}

\item{n_column}{\code{character} (default \code{"n.obs"}) Name of the column containing the number of observations.}

\item{threshold}{\code{numeric} (default \code{1.2}) Threshold to be plotted. Cohorts with a lambda above 
this threshold will be labeled to identify them easily.}
}
\description{
Lambda-N plot
}
\details{
Plot to reveal issues with population stratification. Population stratification can 
either inflate or deflate association P values and can be grasped by the genomic control (GC) 
inflation factor (lambdaGC). Studies with lambdaGC > 1.2 should be revised.
}
