% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFeature.R
\name{lmFeature}
\alias{lmFeature}
\title{Fits a LM to assess association between the features (outcome) and grouping variable (e.g case/control, condition, ...)}
\usage{
lmFeature(feature, vars, Set, cellCountsAdjust, datasources)
}
\arguments{
\item{datasources}{}
}
\value{
a vector with effect estimates, standard error and associated p-value
}
\description{
To be supplied
}
\author{
Gonzalez, JR.
}
