% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusZoom.R
\name{LocusZoom}
\alias{LocusZoom}
\title{LocusZoom}
\usage{
LocusZoom(
  gwas_results,
  range = 5e+05,
  position_zoom = NULL,
  GRCh = 37,
  pvalue = "p.value",
  chromosome = "chr",
  position = "pos"
)
}
\arguments{
\item{gwas_results}{\code{data.frame} With atleast three columns: Chromosome, position and p-value}

\item{range}{\code{numeric} (default \code{5e+05}) Range of the plot in genome positions.}

\item{position_zoom}{\code{numeric} (default \code{NULL}) Center position of the plot if supplied. If \code{NULL}, 
the center position is the lowerst p-value of the \code{gwas_results} table.}

\item{GRCh}{\numeric (default \code{37}) GTCh version to use when querying biomaRt.}

\item{pvalue}{\character (default \code{"p.value"}) Column name of \code{gwas_results} that contains the p-values.}

\item{chromosome}{\character (default \code{"chr"}) Column name of \code{gwas_results} that contains 
the chromosome names.}

\item{position}{\character (default \code{"pos"}) Column name of \code{gwas_results} that contains the positions.}
}
\value{

}
\description{
LocusZoom
}
