% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Mannhattan Plot}
\usage{
manhattan(
  x,
  featureCol = 1,
  chrCol = 2,
  posCol = 3,
  pvalCol = 6,
  sig = 5e-08,
  sugg = 1e-06,
  threshold = NA,
  hlight = NA,
  col = c("#E2709A", "#CB4577", "#BD215B", "#970F42", "#75002B"),
  ylims = c(1.5, 9),
  title = "Manhattan plot",
  ...
)
}
\arguments{
\item{x}{an object of class dsGWAS}

\item{featureCol}{column having feature name (e.g. SNP, CpG, gene, ..)}

\item{chrCol}{column having chromosome name}

\item{sig}{genome-wide significant level. Default 5x10-8}

\item{sugg}{suggestive significant level. Default 1x10-6}

\item{threshold}{numeric value or NA for no label; label snps with pvalue < threshold. Default is NA}

\item{hlight}{list of features names to highlight. For no SNP higlighting, use NA (default)}

\item{col}{vector listing color/colors desired for manhattan plot. Default c("#E2709A", "#CB4577", "#BD215B", "#970F42", "#75002B")}

\item{ylims}{y limits to be visualized in -log10 scale. Default is c(1.5,9)}

\item{title}{plot title, with quotations (e.g. "My Manhattan Plot")}

\item{...}{additional ggplot arguments (ylim, cex, etc. can be altered within the plot section of the function)}

\item{positionCol}{column having basepair position}
}
\value{
a ggplot object having the Manhattan plot
}
\description{
Performs Mannhattan plot for GWAS, DGE or EWAS analysis
}
\author{
modified from here: https://github.com/pcgoddard/Burchardlab_Tutorials/wiki/GGplot2-Manhattan-Plot-Function
}
