% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPvalues.R
\name{metaPvalues}
\alias{metaPvalues}
\title{Meta-analysis of p-values}
\usage{
metaPvalues(x)
}
\arguments{
\item{x}{a \code{dsOmics} object obtained from \code{ds.limma}, \code{ds.GWAS} or \code{ds.PLINK} functions applied o 2 or more studies}
}
\value{
a matrix with features p-values of each study and its combination
}
\description{
Performs meta-analys of pvalues using the sum of logs method (Fisher's method)
}
\author{
Gonzalez, JR.
}
