% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsLmFeature.R
\name{plot.dsLmFeature}
\alias{plot.dsLmFeature}
\title{Generate a QQ-plot of p-values output from any omic data analysis}
\usage{
\method{plot}{dsLmFeature}(x, ...)
}
\arguments{
\item{x}{a "ds.dsLmFeature" type object containing a column of p-values output from a regression analysis}

\item{...}{All other standard R plot arguments available}
}
\description{
The function produces a QQ-plot of the observed vs expected p-values (with confidence bands)
}
\details{
Creates a QQ-plot from omic data analysis
}
\examples{

}
