% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot the results of a \code{ds.PCA()}}
\usage{
plotPCA(object, group = NULL, xcomp = 1, ycomp = 2, datasources = NULL)
}
\arguments{
\item{group}{\code{character} Grouping variable}

\item{xcomp}{\code{numeric} Principal component to plot on the X axis}

\item{ycomp}{\code{numeric} Principal component to plot on the Y axis}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}

\item{res}{\code{PCASNPS} Object returned from \code{ds.PCASNPS()}}
}
\value{
\code{ggplot} object
}
\description{
Plot the results of a \code{ds.PCA()}
}
