% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCASNPS.R
\name{plotPCASNPS}
\alias{plotPCASNPS}
\title{Plot the results of a \code{ds.PCASNPS()}}
\usage{
plotPCASNPS(res, group = NULL, xcomp = 1, ycomp = 2, datasources = NULL)
}
\arguments{
\item{res}{\code{PCASNPS} Object returned from \code{ds.PCASNPS()}}

\item{xcomp}{\code{numeric} Principal component to plot on the X axis}

\item{ycomp}{\code{numeric} Principal component to plot on the Y axis}
}
\value{
\code{ggplot} object
}
\description{
Plot the results of a \code{ds.PCASNPS()}
}
