% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzPlot.R
\name{pzPlot}
\alias{pzPlot}
\title{P-Z plot}
\usage{
pzPlot(x, beta_column = "Est", se_column = "Est.SE", pval_column = "p.value")
}
\arguments{
\item{x}{\code{data.frame / list of data.frames (output of ds.GWAS)} Single table of results or list of tables 
(default output of \code{ds.GWAS}).}

\item{beta_column}{\code{character} (default \code{"Est"}) Name of the column containing the betas.}

\item{se_column}{\code{character} (default \code{"Est.SE"}) Name of the column containing the SE.}

\item{pval_column}{\code{character} (default \code{"p.value"}) Name of the column containing the p-values.}
}
\value{
A ggplot object
}
\description{
P-Z plot
}
\details{
Plot to reveal issues with beta estimates, standard errors and P values.
The plots compare P values reported in the association GWAS with P values calculated from 
Z-statistics (P.ztest) derived from the reported beta and standard error. A study with no issues 
should show perfect concordance. It will generate a plot per study, where concordance of all the study SNPs 
will be displayed.
}
