% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{Create a quantile-quantile plot with ggplot2.}
\usage{
qqplot(ps, ci = 0.95)
}
\arguments{
\item{ps}{Vector of p-values.}

\item{ci}{Size of the confidence interval, 95\% by default.}
}
\value{
A ggplot2 plot.
}
\description{
Assumptions:

  - Expected P values are uniformly distributed.\cr
  - Confidence intervals assume independence between tests.\cr
    We expect deviations past the confidence intervals if the tests are
    not independent.
    For example, in a genome-wide association study, the genotype at any
    position is correlated to nearby positions. Tests of nearby genotypes
    will result in similar test statistics.
}
\examples{
library(ggplot2)
gg_qqplot(runif(1e2)) + theme_grey(base_size = 24)

}
