% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeGenoData.R
\name{standardizeGenoData}
\alias{standardizeGenoData}
\title{Get standaridization coefficients of Genotype data}
\usage{
standardizeGenoData(genoData, datasources)
}
\arguments{
\item{genoData}{\code{character} Name of the \code{GenotypeData} object on the server}

\item{datasources}{a list of \code{DSConnection-class} objects obtained after login}
}
\value{
\code{list} with means and SD of HWE
}
\description{
Get standaridization coefficients of Genotype data
}
