% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolist.R
\name{tolist}
\alias{tolist}
\title{PLINK arguments in the DataSHIELD required format}
\usage{
tolist(args)
}
\arguments{
\item{args}{PLINK arguments}
}
\value{
...
}
\description{
Writes PLINK commands in the DataSHIELD required format (e.g. remove --)
}
\author{
Yannick Marcon
}
