% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.addPhenoData.R
\name{ds.addPhenoData}
\alias{ds.addPhenoData}
\title{Add Phenotype data to ExpressionSet}
\usage{
ds.addPhenoData(
  x,
  pheno,
  identifier = "ID",
  alternate_eset_id = NULL,
  newobj.name = NULL,
  complete_cases = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x}{\code{character} Name of the ExpressionSet on the study server}

\item{pheno}{\code{character} Name of the data.frame with the new phenotypes on the 
study server}

\item{identifier}{\code{character} (default \code{"ID"}) Name of the ID column on the phenotypes data.frame}

\item{alternate_eset_id}{\code{character} (default \code{NULL}) Alternate ID of the eSet pheno data, by default the rownames 
of the eSet pheno data act as ID, use this argument if the ID to merge the individuals is on a column of the pheno data. 
Input NULL for the standard behaviour of using the rownames of the pheno data as ID.}

\item{newobj.name}{\code{character} (default \code{NULL}) If \code{NULL}, the original ExpressionSet will be overwritten,
otherwise the new ExpressionSet will be assigned to a variable named after this argument}

\item{complete_cases}{\code{bool} (default \code{TRUE}) If \code{TRUE} only the matching individuals 
between the ExpressionSet and the phenotypes table will be included on the resulting ExpressionSet. If 
\code{FALSE} all the individuals on the input ExpressionSet will be on the output ExpressionSet}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Add new phenotype data contained on a data.frame to an ExpressionSet. The ExpressionSet
may or may not already have phenotype data. If the data.frame contains a phenotype already present 
on the ExpressionSet, the server function will throw an exception. The pehnotypes data.frame has to 
contain an ID column which does not need to contain exactly the same individuals as the ExpressionSet, 
only the matching individuals will be updated, no new individuals will be introduced or removed from the 
ExpressionSet
}
