% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.alleleFrequency.R
\name{ds.alleleFrequency}
\alias{ds.alleleFrequency}
\title{Allelic frequency}
\usage{
ds.alleleFrequency(
  genoData,
  type = "combined",
  method = "fast",
  snpBlock = 5000L,
  datasources = NULL
)
}
\arguments{
\item{genoData}{\code{character} Name of the \code{\link{GenotypeData}} object on the server}

\item{type}{\code{character} (default \code{"combined"}) Type of analysis, if (\code{"split"}), the
frequencies will be calculated for each study server. If  (\code{"combined"}) 
a pooled methodology will be performed.}

\item{method}{\code{character} (default \code{"fast"}) If \code{"fast"} an optimized fast method will 
be used; if \code{"slow"} the function \code{GWASTools::alleleFrequency} will be used, this is notably 
slower.}

\item{snpBlock}{\code{integer} (default \code{5000L}) number of SNPs to read at each iteration, 
tune this parameter to improve performance. This argument is only used when \code{method = "fast"}}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
A \code{data frame} with a row for each SNP. Columns "M" for males, "F" for females, and "all" 
for all scans give frequencies of the A allele. Sample size for males, females, and all is returned as 
"n.M", "n.F", and "n", respectively. "MAF" is the minor allele frequency over all scans.
}
\description{
Calculates the frequency of the A allele on a server side GenotypeData object.
}
\details{
(From alleleFrequency documentation): Counts male heterozygotes on the X and Y chromosomes
 as missing values, and any genotype for females on the Y chromosome as missing values. A "sex" 
 variable must be present in the scan annotation slot of genoData.
 Samples with missing sex are included in the allele counts for "all" and "MAF" for autosomes, 
 but not for sex chromosomes.
}
