% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.alleleFrequency.R
\name{ds.alleleFrequency}
\alias{ds.alleleFrequency}
\title{Title}
\usage{
ds.alleleFrequency(genoData, datasources = NULL)
}
\arguments{
\item{genoData}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
