% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.alleleFrequency.R
\name{ds.alleleFrequency}
\alias{ds.alleleFrequency}
\title{Allelic frequency}
\usage{
ds.alleleFrequency(
  genoData,
  type = c("combined", "split"),
  method = c("fast", "slow"),
  snpBlock = 5000L,
  datasources = NULL
)
}
\arguments{
\item{genoData}{\code{character} Name of the \code{\link{GenotypeData}} object on the server}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
A \code{data frame} with a row for each SNP. Columns "M" for males, "F" for females, and "all" 
for all scans give frequencies of the A allele. Sample size for males, females, and all is returned as 
"n.M", "n.F", and "n", respectively. "MAF" is the minor allele frequency over all scans.
}
\description{
Calculates the frequency of the A allele on a server side GenotypeData object.
}
\details{
(From alleleFrequency documentation): Counts male heterozygotes on the X and Y chromosomes
 as missing values, and any genotype for females on the Y chromosome as missing values. A "sex" 
 variable must be present in the scan annotation slot of genoData.
 Samples with missing sex are included in the allele counts for "all" and "MAF" for autosomes, 
 but not for sex chromosomes.
}
