% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.DESeq2.R
\name{ds.DESeq2}
\alias{ds.DESeq2}
\title{Server-side Differential gene expression analysis using DESeq2}
\usage{
ds.DESeq2(
  model,
  set,
  test = "Wald",
  fitType = "parametric",
  sfType = "ratio",
  betaPrior = FALSE,
  reduced = NULL,
  contrast = NULL,
  datasources = NULL
)
}
\arguments{
\item{model}{formula indicating the condition and other covariates to be ajusted}

\item{set}{a \code{RangedSummarizedExperiment}}

\item{test}{"Wald" or "LRT"}

\item{fitType}{"parametric", "local", or "mean"}

\item{sfType}{"ratio", "poscounts", or "iterate"}

\item{betaPrior}{TRUE or FALSE}

\item{reduced}{a reduced formula for \code{test="LRT"}}

\item{contrast}{the comparison to extract from the object to build the results table}

\item{datasources}{specifies the particular 'connection object(s)' to use.}
}
\value{
Data frame containing the results of the differential expression analysis
}
\description{
This function is similar to Bioconductor's
\code{DESeq2} package \code{DESeq} function.
}
\details{
Implementation of Bioconductor's \code{DESeq2} in DataSHIELD 
infraestructure.
}
\author{
Abarrategui Leire, for DataSHIELD development team
}
