% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exactHWE.R
\name{ds.exactHWE}
\alias{ds.exactHWE}
\title{Title}
\usage{
ds.exactHWE(
  genoData,
  geno.counts = TRUE,
  snpStart = NULL,
  snpEnd = NULL,
  block.size = 5000,
  permute = FALSE,
  datasources = NULL
)
}
\arguments{
\item{genoData}{}

\item{geno.counts}{}

\item{snpStart}{}

\item{snpEnd}{}

\item{block.size}{}

\item{permute}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
