% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.fastGWAS.R
\name{ds.fastGWAS}
\alias{ds.fastGWAS}
\title{Title}
\usage{
ds.fastGWAS(genoData, formula, family, datasources = NULL)
}
\arguments{
\item{genoData}{}

\item{formula}{}

\item{family}{}

\item{datasources}{}

\item{x}{}
}
\value{

}
\description{
Title
}
