% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.getSNPSbyGen.R
\name{ds.getSNPSbyGen}
\alias{ds.getSNPSbyGen}
\title{Subset GDS with gene(s)}
\usage{
ds.getSNPSbyGen(gdsFile, genes, name = NULL, datasources = NULL)
}
\arguments{
\item{gdsFile}{\code{character} Name of the GDS on the server}

\item{genes}{\code{character vector} Name of the gene(s) to select (HGCN nomenclature)}

\item{name}{\code{character} (default \code{NULL}) Name of the subsetted GDS, if \code{NULL} 
the name \code{gds_subsetted} will be used.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
This function does not have an output. It creates (or overwrites) a GDS object on the study server.
}
\description{
Subset a server side GDS with the SNPs from selected gene(s)
}
