% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.limma.R
\name{ds.limma}
\alias{ds.limma}
\title{Differential expression using limma}
\usage{
ds.limma(model, eSet, type.data = "microarray", sva = FALSE,
  connections = NULL)
}
\arguments{
\item{model}{formula indicating the condition (left side) and other covariates to be adjusted for 
(i.e. condition ~ covar1 + ... + covar2). The fitted model is: feature ~ condition + covar1 + ... + covarN}

\item{eSet}{name of the DataSHIELD object to which the ExpresionSet has been assigned}

\item{type.data}{optional parameter that allows the user to specify the number of CPU cores to use during}

\item{sva}{logical value}

\item{connections}{....}
}
\description{
Performing differential expression analysis using limma
}
\details{
The function  ...
Outputs a matrix containing ...
}
\examples{

}
