% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.PCA.R
\name{ds.PCA}
\alias{ds.PCA}
\title{Title}
\usage{
ds.PCA(
  genoData,
  snp_subset = TRUE,
  standardize = TRUE,
  snpBlock = 20000L,
  ncomp = 5L,
  datasources = NULL
)
}
\arguments{
\item{genoData}{}

\item{standardize}{}

\item{snpBlock}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
