% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.removeOutliers.R
\name{ds.removeOutliers}
\alias{ds.removeOutliers}
\title{Filter potential CpG outliers}
\usage{
ds.removeOutliers(object, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{object}{\code{character} Name of an \code{ExpressionSet} or \code{SummarizedExperiment / RangedSummarizedExperiment}
on the server side to which perform the filtering}

\item{newobj.name}{\code{character} (default \code{NULL}) If provided, the filtered object will be assigned
to a variable on the server with this name. If \code{NULL}, the input variable will be overwritten.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{

}
\description{
Trimming methylation beta values to remove potential outliers.
Trimming scheme is as follows – trim values beyond the lower and upper outer #fences. These are defined by:
Values < 25th percentile minus 3 times IQRANDValues > 75th percentile plus 3 times IQR
(IQR = interquartile range)
}
