% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.table_gds.R
\name{ds.table_gds}
\alias{ds.table_gds}
\title{Title}
\usage{
ds.table_gds(gds, variable, datasources = NULL)
}
\arguments{
\item{gds}{}

\item{variable}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
