% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{Quantile-quantile plot}
\usage{
qqplot(ps, ci = 0.95)
}
\arguments{
\item{ps}{Vector of p-values.}

\item{ci}{Size of the confidence interval, 95% by default.}
}
\value{
A ggplot2 plot.
}
\description{
Creates a QQ-plot to visualize inflation
}
\examples{
library(ggplot2)
gg_qqplot(runif(1e2)) + theme_grey(base_size = 24)
}
