% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeGenoData.R
\name{standardizeGenoData}
\alias{standardizeGenoData}
\title{Title}
\usage{
standardizeGenoData(genoData, snpBlock, datasources)
}
\arguments{
\item{genoData}{}

\item{snpBlock}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
